+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 
 Develop
 -----------------
+* [ENHANCEMENT] Allow  database store backend to support specification of schema in credentials file
 
 
 0.13.0